automatic naming included
for the file containing all unique representative matrices of all files in a given folder, the name of the folder is automatically included in the name of the file. so no need to modify the code for different folders. the code reads the name of the folder it is found in